refactor exception generating instructions
